Moved EnsureInputService into its own file, code review changes
